Implemented custom StoreAdapter Mock for testing.
This allows us to run all of Ottoman's test against a special
mock.  This is neccessary because the mock couchbase bucket is
unable to handle N1QL queries, but at an Ottoman level we are
able to handle the filters at a higher level (rather than as N1QL).
