+++ lib/ottoman.js
@@ -3,6 +3,8 @@
 var util = require('util');
 var Schema = require('./schema');
 var StoreAdapter = require('./storeadapter');
+var CbStoreAdapter = require('./cbstoreadapter');
+var MockStoreAdapter = require('./mockstoreadapter');
 var ModelInstance = require('./modelinstance');
 var autofns = require('./autofns');
 
@@ -985,6 +987,9 @@ Ottoman.loadAll = ModelInstance.loadAll;
 var ottoman = new Ottoman();
 ottoman.Ottoman = Ottoman;
 ottoman.StoreAdapter = StoreAdapter;
-ottoman.CbStoreAdapter = StoreAdapter.Couchbase;
+ottoman.StoreAdapter.Couchbase = CbStoreAdapter;
+ottoman.StoreAdapter.Mock = MockStoreAdapter;
+ottoman.CbStoreAdapter = ottoman.StoreAdapter.Couchbase;
+ottoman.MockStoreAdapter = ottoman.StoreAdapter.Mock;
 ottoman.Consistency = StoreAdapter.SearchConsistency;
 module.exports = ottoman;
