+++ test/harness.js
@@ -1,53 +1,50 @@
 var couchbase = require('couchbase');
 var ottoman = require('../lib/ottoman.js');
 
-// Open the Couchbase Connection
-var cluster = null;
-if (process.env.CNCSTR !== undefined) {
+// Open a connection
+if (process.env.CNCSTR) {
   cluster = new couchbase.Cluster(process.env.CNCSTR);
+  bucket = cluster.openBucket();
+
+  var seenKeys = [];
+  var _bucketInsert = bucket.insert.bind(bucket);
+  bucket.insert = function(key, value, options, callback) {
+    seenKeys.push(key);
+    return _bucketInsert(key, value, options, callback);
+  };
+  var _bucketUpsert = bucket.upsert.bind(bucket);
+  bucket.upsert = function(key, value, options, callback) {
+    seenKeys.push(key);
+    return _bucketUpsert(key, value, options, callback);
+  };
+  var _bucketReplace = bucket.replace.bind(bucket);
+  bucket.replace = function(key, value, options, callback) {
+    seenKeys.push(key);
+    return _bucketReplace(key, value, options, callback);
+  };
+  after(function(done) {
+    if (seenKeys.length === 0) {
+      return done();
+    }
+
+    var remain = seenKeys.length;
+    for (var i = 0; i < seenKeys.length; ++i) {
+      bucket.remove(seenKeys[i], function () {
+        remain--;
+        if (remain === 0) {
+          seenKeys = [];
+          done();
+        }
+      });
+    }
+  });
+
+  ottoman.bucket = bucket;
 } else {
-  cluster = new couchbase.Mock.Cluster();
+  ottoman.store = new ottoman.MockStoreAdapter();
 }
 
-bucket = cluster.openBucket();
-module.exports.bucket = bucket;
-
-// Hijack storage operations on the bucket to track keys
-var seenKeys = [];
-var _bucketInsert = bucket.insert.bind(bucket);
-bucket.insert = function(key, value, options, callback) {
-  seenKeys.push(key);
-  return _bucketInsert(key, value, options, callback);
-};
-var _bucketUpsert = bucket.upsert.bind(bucket);
-bucket.upsert = function(key, value, options, callback) {
-  seenKeys.push(key);
-  return _bucketUpsert(key, value, options, callback);
-};
-var _bucketReplace = bucket.replace.bind(bucket);
-bucket.replace = function(key, value, options, callback) {
-  seenKeys.push(key);
-  return _bucketReplace(key, value, options, callback);
-};
-after(function(done) {
-  if (seenKeys.length === 0) {
-    return done();
-  }
-
-  var remain = seenKeys.length;
-  for (var i = 0; i < seenKeys.length; ++i) {
-    bucket.remove(seenKeys[i], function () {
-      remain--;
-      if (remain === 0) {
-        seenKeys = [];
-        done();
-      }
-    });
-  }
-});
-
 // Setup Ottoman
-ottoman.bucket = bucket;
 module.exports.lib = ottoman;
 
 // Some helpers
@@ -77,5 +74,3 @@ function uniqueId(prefix) {
   return prefix + (uniqueIdCounter++);
 }
 module.exports.uniqueId = uniqueId;
-
-module.exports.cbErrors = couchbase.errors;
